requirements.txt
google-genai==0.3.0
pandas==2.2.2
openpyxl==3.1.5
Flask==3.0.3
python-dotenv==1.0.1
tabulate==0.9.0